chore: refine da_config code style
